Ultima Entrega de Proyecto

+++ Fifa/Fifa/Models/CIUDAD.cs
@@ -20,8 +20,8 @@ namespace Fifa.Models
             this.ESTADIO = new HashSet<ESTADIO>();
         }
     
-        public decimal ID_CIUDAD { get; set; }
-        public decimal ID_PAIS { get; set; }
+        public int ID_CIUDAD { get; set; }
+        public int ID_PAIS { get; set; }
         public string NOMBRE { get; set; }
     
         public virtual PAIS PAIS { get; set; }
